Implemented QueryProcessor separately from Ranker
