+++ app/models/search.rb
@@ -1,7 +1,7 @@
 class Search
 
   def self.query(user_query)
-    Ranker.new(user_query).query
+    QueryProcessor.new(user_query).query
   end
 
 end
